ai/advanced_meta_learning_ai.py
import numpy as np
import torch
import torch.nn as nn
from torch_geometric.nn import GCNConv
from transformers import BertModel, BertTokenizer
from qiskit_aer import Aer
from qiskit.primitives import Sampler
from qiskit_algorithms import QAOA
from qiskit_optimization import QuadraticProgram
from qiskit_algorithms.optimizers import COBYLA
from pyro.infer import SVI, Trace_ELBO
import pyro
import pyro.distributions as dist
from ai.explainability_module import ExplainabilityModule

from ai.memory_bank import MemoryBank
from nas_nn import NASNeuralNetwork
from ai.maml import maml_step
from ai.quantum_optimizer import QuantumOptimizer

class AdvancedMetaLearningAI:
    def __init__(self, coder):
        self.state = self.initialize_state()
        self.knowledge_base = self.create_knowledge_base()
        self.model = NASNeuralNetwork()
        self.memory_buffer = []
        self.buffer_max_size = 100  # New buffer size limit
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
        self.bert_model = BertModel.from_pretrained('bert-base-uncased')
        self.backend = Aer.get_backend('qasm_simulator')  # Updated backend usage
        self.optimizer = COBYLA()
        self.sampler = Sampler()
        self.coder = coder
        self.maml_model = MAMLModel(input_size=10, hidden_size=40, output_size=10)
        self.maml_optimizer = torch.optim.Adam(self.maml_model.parameters(), lr=0.001)
        self.quantum_optimizer = QuantumOptimizer(num_qubits=5)
        self.meta_learners = {
            'MAML': MAML(),
            'Reptile': Reptile(),
            'PrototypicalNetworks': PrototypicalNetworks()
        }
        
    def initialize_state(self):
        return np.random.rand(10)

    def create_knowledge_base(self):
        return np.random.rand(10)

    def spark_intelligence(self):
        potential_states = self.evolve_states(self.state)
        self.knowledge_base = self.optimize(potential_states)
        insight = self.instant_insight(self.knowledge_base)
        self.state = self.collapse_to_superintelligence(insight)

    def evolve_states(self, state):
        # Define an optimization problem
        problem = QuadraticProgram()
        for i in range(len(state)):
            problem.binary_var(f"x_{i}")
        
        # Objective: minimize the difference between variables and desired state
        linear = {f"x_{i}": -state[i] for i in range(len(state))}
        problem.minimize(linear=linear)
        
        # Initialize QAOA
        qaoa = QAOA(optimizer=self.optimizer, quantum_instance=self.backend)
        
        # Convert problem to Ising Hamiltonian
        operator, offset = problem.to_ising()
        
        # Execute QAOA
        result = qaoa.compute_minimum_eigenvalue(operator)
        
        # Extract the solution
        x_opt = problem.interpret(result)
        potential_states = np.array([list(x_opt.values())], dtype=np.float32)
        
        return potential_states

    def optimize_with_meta_learner(self, potential_states):
        # Determine the best meta-learner for the task
        # Placeholder for selection logic
        selected_learner = self.meta_learners['MAML']
        # Perform meta-training
        selected_learner.meta_train(self.model, tasks)
    
    def optimize_with_quantum(self, potential_states):
        def optimize(self, potential_states):
        # Define cost operator based on potential_states
        # Placeholder for actual cost operator
        cost_operator = PauliSumOp.from_list([("Z" * 5, 1.0)])
        result = self.quantum_optimizer.optimize(cost_operator)
        # Use result to update knowledge base
        optimized_state = result.x
        self.knowledge_base = optimized_state[:10]  # Update knowledge base
    
    def optimize(self, potential_states):
        # Ensure potential_states is 2D (each state is a row)
        if len(potential_states.shape) == 1:
            potential_states = np.expand_dims(potential_states, axis=0)  # Convert 1D to 2D if necessary

        # Create a PyTorch tensor from potential_states
        X = torch.tensor(potential_states, dtype=torch.float32)  # Now, X should be 2D

        # Summing along axis 1
        y = torch.tensor(np.sum(X.numpy(), axis=1), dtype=torch.float32)

        if len(self.memory_buffer) > 0:
            X_memory, y_memory = zip(*self.memory_buffer)
            X_memory = torch.tensor(np.vstack(X_memory), dtype=torch.float32)
            
            # Handle the case where y_memory contains scalar values or arrays
            y_memory = torch.tensor([y if np.isscalar(y) else y.item() for y in y_memory], dtype=torch.float32)
            
            X = torch.cat((X, X_memory), dim=0)
            y = torch.cat((y, y_memory), dim=0)

        # Limit buffer size to prevent memory issues
        if len(self.memory_buffer) >= self.buffer_max_size:
            self.memory_buffer = self.memory_buffer[-self.buffer_max_size:]

        # Pyro optimizer
        pyro_optimizer = pyro.optim.Adam({"lr": 0.001})

        # Use the Pyro optimizer in SVI
        svi = SVI(self.model, self.guide, pyro_optimizer, loss=Trace_ELBO())

        for _ in range(100):  # Training step
            loss = svi.step(X, y)

        # Ensure we're returning a consistent shape (e.g., always 1D array of size 10)
        result = self.model(X).detach().numpy()
        if result.ndim > 1:
            result = result.mean(axis=0)  # Average across all samples if more than 1D
        result = result[:10]  # Ensure we always return 10 elements

        # Ensure we're appending numpy arrays, not tensors
        self.memory_buffer.append((X[-1].detach().numpy(), y[-1].detach().numpy()))  # Updated buffer
        
        tasks = self.create_meta_tasks(X, y)
        for epoch in range(10):
            loss = maml_step(self.maml_model, self.maml_optimizer, tasks)
        # Use the adapted model to predict
        result = self.maml_model(X).detach().numpy()
        
        return result
    
    def create_meta_tasks(self, X, y):
        # Create tasks for MAML
        # Placeholder implementation
        tasks = []
        for i in range(len(X)):
            x_train = X[i:i+1]
            y_train = y[i:i+1]
            x_val = X[i+1:i+2] if i+1 < len(X) else X[0:1]
            y_val = y[i+1:i+2] if i+1 < len(y) else y[0:1]
            tasks.append((x_train, y_train, x_val, y_val))
        return tasks
    
    def guide(self, X, y):
        pyro.module("model", self.model)
        with pyro.plate("data", len(X)):
            prior_loc = torch.zeros_like(y)  # Ensure the shape matches y
            prior_scale = torch.ones_like(y)
            pyro.sample("obs", dist.Normal(prior_loc, prior_scale), obs=y)

    def instant_insight(self, knowledge_base):
        # Insight mechanism using BERT
        inputs = self.tokenizer("your text", return_tensors='pt')
        outputs = self.bert_model(**inputs)
        # Returning averaged BERT output for use in decision making
        return outputs.last_hidden_state.mean().item()

    def collapse_to_superintelligence(self, insight):
        class KnowledgeGraph(nn.Module):
            def __init__(self):
                super(KnowledgeGraph, self).__init__()
                self.conv1 = GCNConv(10, 64)
                self.conv2 = GCNConv(64, 10)

            def forward(self, x, edge_index):
                x = torch.relu(self.conv1(x, edge_index))
                return self.conv2(x, edge_index)

        # Adjusting edge_index to have valid node indices
        num_nodes = 10  # Assuming you want 10 nodes for this graph
        edge_index = torch.tensor([[0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
                                [1, 2, 3, 4, 5, 6, 7, 8, 9, 0]], dtype=torch.long)

        # Ensure that x has the correct number of nodes to match edge_index
        x = torch.tensor(self.state, dtype=torch.float32).unsqueeze(0).repeat(num_nodes, 1)  # Creating 10 node features
        knowledge_graph = KnowledgeGraph()
        graph_output = knowledge_graph(x, edge_index)

        # Combining BERT insight with GCN-based state
        attention_weights = torch.softmax(torch.tensor([insight] * len(self.state)), dim=0)

        # Use matrix multiplication to combine the attention weights and graph output
        combined_state = torch.matmul(attention_weights, graph_output).detach().numpy()

        return combined_state

    # Improved federated learning with local model updates
    def federated_learning_update(self, local_updates):
        # Ensure all local updates have the same shape
        local_updates = [update[:10] if len(update) >= 10 else np.pad(update, (0, 10 - len(update))) for update in local_updates]
        global_update = np.mean(local_updates, axis=0)
        self.state += global_update

    def explain_decision(self):
        model_input = torch.tensor(self.state, dtype=torch.float32).unsqueeze(0)
        explain_module = ExplainabilityModule(self.model)
        attributions = explain_module.explain(model_input, method='integrated_gradients')
        # Visualize or process attributions as needed
        return attributions