requirements.txt
cma
numpy
pyro-ppl
qiskit
qiskit_aer
qiskit_optimization
qiskit-algorithms
shap
torch
torch_geometric
transformers
matplotlib
selenium
gliner
accelerate
coloredlogs
einops
huggingface_hub
termcolor
pycapnp
astor